Clean up and alighment with Csharp Monero lib

+++ Plugins/Monero/Controllers/MoneroLikeStoreController.cs
@@ -301,7 +301,7 @@ namespace BTCPayServer.Plugins.Monero.Controllers
             [Display(Name = "Consider the invoice settled when the payment transaction …")]
             public MoneroLikeSettlementThresholdChoice SettlementConfirmationThresholdChoice { get; set; }
             [Display(Name = "Required Confirmations"), Range(0, 100)]
-            public long? CustomSettlementConfirmationThreshold { get; set; }
+            public int? CustomSettlementConfirmationThreshold { get; set; }
 
             public IEnumerable<ValidationResult> Validate(ValidationContext validationContext)
             {
