+++ Plugins/Monero/Payments/MoneroLikeOnChainPaymentMethodDetails.cs
@@ -2,8 +2,8 @@ namespace BTCPayServer.Plugins.Monero.Payments
 {
     public class MoneroLikeOnChainPaymentMethodDetails
     {
-        public uint AccountIndex { get; set; }
+        public long AccountIndex { get; set; }
         public uint AddressIndex { get; set; }
-        public long? InvoiceSettledConfirmationThreshold { get; set; }
+        public int? InvoiceSettledConfirmationThreshold { get; set; }
     }
 }
