+++ Plugins/Monero/Payments/MoneroPaymentPromptDetails.cs
@@ -3,6 +3,6 @@ namespace BTCPayServer.Plugins.Monero.Payments
     public class MoneroPaymentPromptDetails
     {
         public long AccountIndex { get; set; }
-        public long? InvoiceSettledConfirmationThreshold { get; set; }
+        public int? InvoiceSettledConfirmationThreshold { get; set; }
     }
 }
